Add loop test
