+++ internal/ffmpeg/buffer_test.go
@@ -1,35 +1,11 @@
 package ffmpeg_test
 
 import (
-	"errors"
 	"testing"
 	"time"
 
 	"github.com/dianelooney/gggv/internal/ffmpeg"
 )
-
-type mockSlow struct {
-}
-
-func (r *mockSlow) Read() (ffmpeg.Frame, error) {
-	time.Sleep(time.Second)
-	return ffmpeg.Frame{}, nil
-}
-
-type mockFast struct {
-	called int
-}
-
-func (r *mockFast) Read() (ffmpeg.Frame, error) {
-	r.called++
-	return ffmpeg.Frame{}, nil
-}
-
-type mockErr struct{}
-
-func (r mockErr) Read() (ffmpeg.Frame, error) {
-	return ffmpeg.Frame{}, errors.New("something went wrong")
-}
 
 func TestBufferedReader(t *testing.T) {
 	t.Run("It waits for the underlying reader", func(t *testing.T) {
